requirements.txt
beautifulsoup4
bson
facebook_scraper==0.2.56
Flask==2.1.1
Flask_Cors==3.0.10
google==3.0.0
googletrans==4.0.0rc1
gunicorn==20.1.0
nltk==3.3
numpy==1.21.5
opencv-python-headless==4.2.0.32
pillow==9.1.0
pymongo==4.1.1
pymongo[srv]
pytesseract==0.3.9
pytwitterscraper==1.3.5
requests==2.25.0
scikit_learn==1.0.2
twython==3.9.1
tensorflow==1.15.3
urllib3==1.26.2
